refactor(serializer): move _serializer.py → utils/serialization.py
Mirrors sqlspec's layout (`sqlspec/utils/serializers/_schema.py`) and
frees up the top-level `litestar_mcp/` namespace. `utils.py` becomes
the `utils/` package's `__init__.py`; the serializer module lives at
`utils/serialization.py`.

Pure rename + one import update in `litestar_mcp.typing` and the
serializer test module. Public API unchanged — `from litestar_mcp.typing
import schema_dump` still works.
